Add Death logic for Enemy
