+++ Assets/Prefabs/Enemy.prefab
@@ -79,6 +79,7 @@ GameObject:
   - component: {fileID: 6178350791534739370}
   - component: {fileID: 7632970438909478808}
   - component: {fileID: 8218128473041881738}
+  - component: {fileID: 8606495095698628831}
   m_Layer: 0
   m_Name: Enemy
   m_TagString: Untagged
@@ -178,3 +179,15 @@ CapsuleCollider:
   m_Height: 2
   m_Direction: 1
   m_Center: {x: 0.000000059604645, y: 0, z: -0.00000008940697}
+--- !u!114 &8606495095698628831
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8607934531943104907}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: dc52bbd2bb1e46fea43eac81e894bdf0, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
