--- Assets/Scripts/DamageArea.cs
@@ -1,17 +1,0 @@
-using Enemies;
-using UnityEngine;
-
-public class DamageArea : MonoBehaviour
-{
-    [HideInInspector] public EnemyBase enemy;
-
-    void Awake()
-    {
-        enemy = GetComponentInParent<EnemyBase>();
-    }
-
-    public void OnDamage()
-    {
-        enemy.Damage();
-    }
-}
