+++ Assets/Scripts/Enemies/DamageArea.cs
@@ -1,0 +1,19 @@
+using UnityEngine;
+
+namespace Enemies
+{
+    public class DamageArea : MonoBehaviour
+    {
+        [HideInInspector] public EnemyBase enemy;
+
+        void Awake()
+        {
+            enemy = GetComponentInParent<EnemyBase>();
+        }
+
+        public void OnDamage()
+        {
+            enemy.Damage();
+        }
+    }
+}
