+++ Assets/Scripts/Enemies/NormalEnemy.cs
@@ -1,3 +1,5 @@
+using UnityEngine;
+
 namespace Enemies
 {
     public class NormalEnemy : EnemyBase
@@ -8,12 +10,18 @@ namespace Enemies
         // 시작할 때 HP 초기화
         private void Start()
         {
-            hp = 10f;
+            hp = 5f;
         }
 
         public override void Damage()
         {
+            Debug.Log($"{gameObject.name} 데미지 : HP {hp}");
             hp -= 1f;
+            if (hp <= 0f)
+            {
+                // 사망
+                Destroy(gameObject);
+            }
         }
     }
 }
